devel version 3.1.7~pre1
git-svn-id: https://svn.code.sf.net/p/xsnow/code/trunk@375 79837d8a-40e5-47fd-869e-05d31bbd7fe9

+++ xsnow-devel/src/vroot.h
@@ -81,7 +81,7 @@ VirtualRootWindowOfScreen(screen)
       __SWM_VROOT = XInternAtom(dpy, "__SWM_VROOT", False);
       if (XQueryTree(dpy, root, &rootReturn, &parentReturn,
 	       &children, &numChildren)) {
-	 for (i = 0; i < numChildren; i++) {
+	 for (i = 0; i < (int)numChildren; i++) {
 	    Atom actual_type;
 	    int actual_format;
 	    unsigned long nitems, bytesafter;
